+++ app/src/js/components/Collections/overview.js
@@ -20,10 +20,13 @@ import {
   collectionNameVersion,
   getCollectionId,
   lastUpdated,
-  tally,
 } from '../../utils/format';
 import pageSizeOptions from '../../utils/page-size';
 import statusOptions from '../../utils/status';
+import {
+  getPersistentQueryParams,
+  historyPushWithQueryParams,
+} from '../../utils/url-helper';
 import isEqual from 'lodash.isequal';
 import {
   reingestAction,
@@ -95,7 +98,7 @@ class CollectionOverview extends React.Component {
 
   changeCollection(_, collectionId) {
     const { name, version } = collectionNameVersion(collectionId);
-    this.props.history.push(`/collections/collection/${name}/${version}`);
+    historyPushWithQueryParams(`/collections/collection/${name}/${version}`);
   }
 
   generateBulkActions() {
@@ -115,8 +118,10 @@ class CollectionOverview extends React.Component {
   }
 
   generateQuery() {
+    const { match, queryParams } = this.props;
     return {
-      collectionId: getCollectionId(this.props.match.params),
+      ...queryParams,
+      collectionId: getCollectionId(match.params),
     };
   }
 
@@ -126,11 +131,11 @@ class CollectionOverview extends React.Component {
   }
 
   navigateBack() {
-    this.props.history.push('/collections/all');
+    historyPushWithQueryParams('/collections/all');
   }
 
   gotoGranules() {
-    this.props.history.push('/granules');
+    historyPushWithQueryParams('/granules');
   }
 
   errors() {
@@ -140,17 +145,6 @@ class CollectionOverview extends React.Component {
       get(this.props.collections.map, [collectionId, 'error']),
       get(this.props.collections.deleted, [collectionId, 'error']),
     ].filter(Boolean);
-  }
-
-  renderOverview(record) {
-    const data = get(record, 'data', {});
-    const stats = get(data, 'stats', {});
-    const overview = [
-      [tally(stats.completed), strings.granules_completed],
-      [tally(stats.failed), strings.granules_failed],
-      [tally(stats.running), strings.granules_running],
-    ];
-    return <Overview items={overview} inflight={record.inflight} />;
   }
 
   renderDeleteButton() {
@@ -182,7 +176,6 @@ class CollectionOverview extends React.Component {
       collections,
       granules: { list },
     } = this.props;
-
     const collectionName = params.name;
     const collectionVersion = params.version;
     const collectionId = getCollectionId(params);
@@ -191,9 +184,6 @@ class CollectionOverview extends React.Component {
       (id1, id2) => id1.localeCompare(id2, 'en', { sensitivity: 'base' })
     );
     const record = collections.map[collectionId];
-
-    // create the overview boxes
-    const overview = record ? this.renderOverview(record) : <div></div>;
 
     return (
       <div className="page__component">
@@ -233,13 +223,14 @@ class CollectionOverview extends React.Component {
               <li>
                 <Link
                   className="button button--copy button--small button--green"
-                  to={{
+                  to={(location) => ({
                     pathname: '/collections/add',
+                    search: getPersistentQueryParams(location),
                     state: {
                       name: collectionName,
                       version: collectionVersion,
                     },
-                  }}
+                  })}
                 >
                   Copy
                 </Link>
@@ -247,7 +238,10 @@ class CollectionOverview extends React.Component {
               <li>
                 <Link
                   className="button button--edit button--small button--green"
-                  to={`/collections/edit/${collectionName}/${collectionVersion}`}
+                  to={(location) => ({
+                    pathname: `/collections/edit/${collectionName}/${collectionVersion}`,
+                    search: getPersistentQueryParams(location),
+                  })}
                 >
                   Edit
                 </Link>
@@ -265,7 +259,7 @@ class CollectionOverview extends React.Component {
               Granule Metrics
             </h2>
           </div>
-          {overview}
+          {record && <Overview type='granules' params={{ collectionId }} inflight={record.inflight} />}
         </section>
         <section className="page__section">
           <div className="heading__wrapper--border">
@@ -277,7 +271,10 @@ class CollectionOverview extends React.Component {
             </h2>
             <Link
               className="button button--small button__goto button--green form-group__element--right"
-              to={`/collections/collection/${collectionName}/${collectionVersion}/granules`}
+              to={(location) => ({
+                pathname: `/collections/collection/${collectionName}/${collectionVersion}/granules`,
+                search: getPersistentQueryParams(location),
+              })}
             >
               {strings.view_all_granules}
             </Link>
@@ -335,9 +332,8 @@ CollectionOverview.propTypes = {
   datepicker: PropTypes.object,
   dispatch: PropTypes.func,
   granules: PropTypes.object,
-  history: PropTypes.object,
   match: PropTypes.object,
-  router: PropTypes.object,
+  queryParams: PropTypes.object,
 };
 
 export default withRouter(