+++ app/src/js/components/Executions/execution-events.js
@@ -20,6 +20,7 @@ import Search from '../Search/search';
 
 import { getEventDetails } from './execution-graph-utils';
 import SortableTable from '../SortableTable/SortableTable';
+import { historyPushWithQueryParams } from '../../utils/url-helper';
 
 class ExecutionEvents extends React.Component {
   constructor () {
@@ -30,16 +31,16 @@ class ExecutionEvents extends React.Component {
   }
 
   componentDidMount () {
-    const { dispatch } = this.props;
-    const { executionArn } = this.props.match.params;
+    const { dispatch, match } = this.props;
+    const { executionArn } = match.params;
     dispatch(getExecutionStatus(executionArn));
     dispatch(getCumulusInstanceMetadata());
   }
 
   componentDidUpdate (prevProps) {
-    const { dispatch } = this.props;
-    const { executionArn } = this.props.match.params;
-    const { search } = this.props.location;
+    const { dispatch, match, location } = this.props;
+    const { executionArn } = match.params;
+    const { search } = location;
     const { search: prevSearch } = prevProps.location;
     if (search !== prevSearch) {
       dispatch(getExecutionStatus(executionArn));
@@ -68,8 +69,7 @@ class ExecutionEvents extends React.Component {
   }
 
   navigateBack () {
-    const { history } = this.props;
-    history.push('/executions');
+    historyPushWithQueryParams('/executions');
   }
 
   errors () {
@@ -150,7 +150,6 @@ ExecutionEvents.propTypes = {
   match: PropTypes.object,
   dispatch: PropTypes.func,
   location: PropTypes.object,
-  history: PropTypes.object
 };
 
 ExecutionEvents.displayName = 'Execution Events';