+++ app/src/js/components/Granules/overview.js
@@ -4,6 +4,7 @@ import { Helmet } from 'react-helmet';
 import PropTypes from 'prop-types';
 import { withRouter } from 'react-router-dom';
 import { connect } from 'react-redux';
+import withQueryParams from 'react-router-query-params';
 import {
   searchGranules,
   clearGranulesSearch,
@@ -17,7 +18,7 @@ import {
   getGranuleCSV
 } from '../../actions';
 import { get } from 'object-path';
-import { lastUpdated, tally, displayCase } from '../../utils/format';
+import { lastUpdated, tally } from '../../utils/format';
 import {
   tableColumns,
   simpleDropdownOption,
@@ -81,7 +82,8 @@ class GranulesOverview extends React.Component {
   }
 
   generateQuery () {
-    return {};
+    const { queryParams } = this.props;
+    return { ...queryParams };
   }
 
   generateBulkActions () {
@@ -138,11 +140,10 @@ class GranulesOverview extends React.Component {
   }
 
   render () {
-    const { stats, granules, dispatch } = this.props;
-    const { list, dropdowns } = granules;
+    const { collections, dispatch, granules } = this.props;
+    const { list } = granules;
+    const { dropdowns } = collections;
     const { count, queriedAt } = list.meta;
-    const statsCount = get(stats, 'count.data.granules.count', []);
-    const overviewItems = statsCount.map(d => [tally(d.count), displayCase(d.key)]);
     return (
       <div className='page__component'>
         <Helmet>
@@ -155,7 +156,7 @@ class GranulesOverview extends React.Component {
           <div className='page__section__header'>
             <h1 className='heading--large heading--shared-content with-description '>{strings.granule_overview}</h1>
             {lastUpdated(queriedAt)}
-            <Overview items={overviewItems} inflight={false} />
+            <Overview type='granules' inflight={false} />
           </div>
         </section>
         <section className='page__section'>
@@ -178,7 +179,7 @@ class GranulesOverview extends React.Component {
             <ListFilters>
               <Dropdown
                 getOptions={getOptionsCollectionName}
-                options={get(dropdowns, ['collectionName', 'options'])}
+                options={get(dropdowns, ['collectionName', 'options']) || []}
                 action={filterGranules}
                 clear={clearGranulesFilter}
                 paramKey='collectionId'
@@ -223,20 +224,21 @@ class GranulesOverview extends React.Component {
 }
 
 GranulesOverview.propTypes = {
-  granules: PropTypes.object,
-  stats: PropTypes.object,
-  dispatch: PropTypes.func,
-  workflowOptions: PropTypes.array,
+  collections: PropTypes.object,
   config: PropTypes.object,
-  granuleCSV: PropTypes.object
+  dispatch: PropTypes.func,
+  granuleCSV: PropTypes.object,
+  granules: PropTypes.object,
+  queryParams: PropTypes.object,
+  workflowOptions: PropTypes.array,
 };
 
 export { GranulesOverview };
 
-export default withRouter(connect(state => ({
-  stats: state.stats,
-  workflowOptions: workflowOptionNames(state),
-  granules: state.granules,
+export default withRouter(withQueryParams()(connect(state => ({
+  collections: state.collections,
   config: state.config,
-  granuleCSV: state.granuleCSV
-}))(GranulesOverview));
+  granuleCSV: state.granuleCSV,
+  granules: state.granules,
+  workflowOptions: workflowOptionNames(state),
+}))(GranulesOverview)));