+++ app/src/js/components/Rules/add.js
@@ -11,7 +11,7 @@ import {
   listCollections
 } from '../../actions';
 import AddRecord from '../Add/add';
-import { getCollectionId, collectionNameVersion } from '../../utils/format';
+import { getCollectionId, nullValue, collectionNameVersion } from '../../utils/format';
 
 /**
  * Converts the Collection ID string associated with the `collection` property
@@ -44,7 +44,10 @@ import { getCollectionId, collectionNameVersion } from '../../utils/format';
  *    function does not perform any validation)
  */
 const validate = (rule) => {
-  rule.collection = collectionNameVersion(rule.collection);
+  const ruleCollection = collectionNameVersion(rule.collection);
+  if (ruleCollection !== nullValue) {
+    rule.collection = ruleCollection;
+  }
   return true;
 };
 
@@ -108,17 +111,17 @@ const AddRule = ({
     }
   }, [name, rulesMap, isCopy]);
 
-  const dispatched = ({ list }) => {
-    const { inflight, meta, error } = list || {};
-    const { queriedAt } = meta || {};
-    return error || inflight || queriedAt !== undefined;
-  };
+  useEffect(() => {
+    dispatch(listCollections({ listAll: true, getMMT: false }));
+  }, [dispatch]);
 
   useEffect(() => {
-    if (!dispatched(collections)) dispatch(listCollections({ listAll: true, getMMT: false }));
-    if (!dispatched(providers)) dispatch(listProviders({ listAll: true }));
-    if (!dispatched(workflows)) dispatch(listWorkflows());
-  });
+    dispatch(listProviders());
+  }, [dispatch]);
+
+  useEffect(() => {
+    dispatch(listWorkflows());
+  }, [dispatch]);
 
   useEffect(() => {
     setEnums({